Update Affiliation CRUD to support multipart files

+++ src/main/java/com/ensiasit/projectx/services/AffiliationServiceImpl.java
@@ -1,90 +1,85 @@
 package com.ensiasit.projectx.services;
 
-import com.ensiasit.projectx.dto.AffiliationDto;
+import com.ensiasit.projectx.dto.AffiliationRequest;
+import com.ensiasit.projectx.dto.AffiliationResponse;
+import com.ensiasit.projectx.exceptions.ForbiddenException;
 import com.ensiasit.projectx.exceptions.NotFoundException;
+import com.ensiasit.projectx.exceptions.ServerErrorException;
+import com.ensiasit.projectx.mappers.AffiliationMapper;
 import com.ensiasit.projectx.models.Affiliation;
 import com.ensiasit.projectx.repositories.AffiliationRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class AffiliationServiceImpl implements AffiliationService {
-
     private final AffiliationRepository affiliationRepository;
+    private final AffiliationMapper affiliationMapper;
+    private final AdminService adminService;
 
     @Override
-    public List<AffiliationDto> getAll() {
-
+    public List<AffiliationResponse> getAll() {
         return affiliationRepository.findAll().stream()
-                .map(affiliation -> AffiliationDto.builder()
-                        .id(affiliation.getId())
-                        .name(affiliation.getName())
-                        .country(affiliation.getCountry())
-                        .logo(affiliation.getLogo())
-                        .build())
+                .map(affiliationMapper::toDto)
                 .toList();
     }
 
     @Override
-    public AffiliationDto createAffiliation(AffiliationDto affiliationDto) {
-        Affiliation affiliation = affiliationRepository.save(Affiliation.builder()
-                .name(affiliationDto.getName())
-                .country(affiliationDto.getCountry())
-                .logo(affiliationDto.getLogo())
-                .build());
+    public AffiliationResponse createAffiliation(String userEmail, AffiliationRequest affiliationDto) {
+        if (!adminService.isAdmin(userEmail)) {
+            throw new ForbiddenException("User is not admin");
+        }
 
-        return AffiliationDto.builder()
-                .id(affiliation.getId())
-                .name(affiliation.getName())
-                .country(affiliation.getCountry())
-                .logo(affiliation.getLogo())
-                .build();
+        Affiliation affiliation = affiliationRepository.save(affiliationMapper.fromDto(affiliationDto));
+
+        return affiliationMapper.toDto(affiliation);
     }
 
     @Override
-    public AffiliationDto getAffiliation(long id) {
+    public AffiliationResponse getAffiliation(long id) {
         Affiliation affiliation = extractAffiliation(id);
 
-        return AffiliationDto.builder()
-                .id(affiliation.getId())
-                .name(affiliation.getName())
-                .country(affiliation.getCountry())
-                .logo(affiliation.getLogo())
-                .build();
+        return affiliationMapper.toDto(affiliation);
     }
 
     @Override
-    public AffiliationDto deleteAffiliation(long id) {
+    public AffiliationResponse deleteAffiliation(String userEmail, long id) {
+        if (!adminService.isAdmin(userEmail)) {
+            throw new ForbiddenException("User is not admin");
+        }
+
         Affiliation affiliation = extractAffiliation(id);
+
         affiliationRepository.deleteById(id);
 
-        return AffiliationDto.builder()
-                .id(affiliation.getId())
-                .name(affiliation.getName())
-                .country(affiliation.getCountry())
-                .logo(affiliation.getLogo())
-                .build();
+        return affiliationMapper.toDto(affiliation);
     }
 
     @Override
-    public AffiliationDto updateAffiliation(long id, AffiliationDto payload) {
+    public AffiliationResponse updateAffiliation(String userEmail, long id, AffiliationRequest payload) {
+        if (!adminService.isAdmin(userEmail)) {
+            throw new ForbiddenException("User is not admin");
+        }
+
         Affiliation affiliation = extractAffiliation(id);
 
         affiliation.setName(payload.getName());
         affiliation.setCountry(payload.getCountry());
-        affiliation.setLogo(payload.getLogo());
+
+        try {
+            affiliation.setLogo(payload.getLogo().getBytes());
+        } catch (IOException e) {
+            throw new ServerErrorException("Could not read file content.");
+        }
+
         affiliationRepository.save(affiliation);
 
-        return AffiliationDto.builder()
-                .id(affiliation.getId())
-                .name(affiliation.getName())
-                .country(affiliation.getCountry())
-                .logo(affiliation.getLogo())
-                .build();
+        return affiliationMapper.toDto(affiliation);
     }
 
     private Affiliation extractAffiliation(long id) {
